Added feedback from @kemar
